tidy up mock data
