+++ lambdas/tests/unit/models/test_statistics_models.py
@@ -1,7 +1,7 @@
 import pydantic
 import pytest
 from models.statistics import ApplicationData, RecordStoreData, load_from_dynamodb_items
-from tests.unit.helpers.data.statistic.statistic_data import (
+from tests.unit.helpers.data.statistic.mock_statistic_data import (
     MOCK_APPLICATION_DATA_1,
     MOCK_APPLICATION_DATA_2,
     MOCK_DYNAMODB_ITEMS,
